refactor: load youtube iframe api at top.
Showing a video and loading youtube api are two different concerns. Load api at
the top of the tree. We can't do much in the app if youtube doesn't load.

In future we might have to care about abstracting away youtube specific details.
Example, when we get to supporting vimeo etc.